# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.AlertRuleForm:
    properties:
      enable:
        type: boolean
      fail_trigger_tasks:
        description: 失败时触发的任务id
        items:
          type: integer
        type: array
      name:
        minLength: 1
        type: string
      notification_group_id:
        type: integer
      recover_trigger_tasks:
        description: 恢复时触发的任务id
        items:
          type: integer
        type: array
      rules:
        items:
          $ref: '#/definitions/model.Rule'
        type: array
      trigger_mode:
        default: 0
        type: integer
    type: object
...
  model.CommonResponse-any:
    properties:
      data: {}
      error:
        type: string
      success:
        type: boolean
    type: object
  model.CommonResponse-array_model_AlertRule:
    properties:
      data:
        items:
          $ref: '#/definitions/model.AlertRule'
        type: array
      error:
        type: string
      success:
        type: boolean
    type: object
...
  model.CommonResponse-uint64:
    properties:
      data:
        type: integer
      error:
        type: string
      success:
        type: boolean
    type: object
  model.CreateFMResponse:
    properties:
      session_id:
        type: string
    type: object
...
  model.Oauth2LoginResponse:
    properties:
      redirect:
        type: string
    type: object
...
host: localhost:8008
info:
  contact:
    email: hi@nai.ba
    name: API Support
    url: http://nezhahq.github.io
  description: Nezha Monitoring API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://nezhahq.github.io
  title: Nezha Monitoring API
  version: "1.0"
paths:
  /alert-rule:
    get:
      description: List Alert rules
      parameters:
      - description: Resource ID
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-array_model_AlertRule'
      security:
      - BearerAuth: []
      summary: List Alert rules
      tags:
      - auth required
    post:
      consumes:
      - application/json
      description: Add Alert Rule
      parameters:
      - description: AlertRuleForm
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.AlertRuleForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-uint64'
      security:
      - BearerAuth: []
      summary: Add Alert Rule
      tags:
      - auth required
  /alert-rule/{id}:
    patch:
      consumes:
      - application/json
      description: Update Alert Rule
      parameters:
      - description: Alert ID
        in: path
        name: id
        required: true
        type: integer
      - description: AlertRuleForm
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.AlertRuleForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: Update Alert Rule
      tags:
      - auth required
  /api/v1/oauth2/{provider}:
    get:
      description: Get Oauth2 Redirect URL
      parameters:
      - description: provider
        in: path
        name: provider
        required: true
        type: string
      - default: 1
        description: type
        enum:
        - 1
        - 2
        in: query
        name: type
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Oauth2LoginResponse'
      summary: Get Oauth2 Redirect URL
  /api/v1/oauth2/{provider}/unbind:
    post:
      consumes:
      - application/json
      description: Unbind Oauth2
      parameters:
      - description: provider
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema: {}
      summary: Unbind Oauth2
  /api/v1/oauth2/callback:
    get:
      consumes:
      - application/json
      description: Oauth2 Callback
      parameters:
      - description: state
        in: query
        name: state
        required: true
        type: string
      - description: code
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      summary: Oauth2 Callback
  /batch-delete/alert-rule:
    post:
      consumes:
      - application/json
      description: Batch delete Alert rules
      parameters:
      - description: id list
        in: body
        name: request
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: Batch delete Alert rules
      tags:
      - auth required
  /batch-delete/cron:
    post:
      consumes:
      - application/json
      description: Batch delete schedule tasks
      parameters:
      - description: id list
        in: body
        name: request
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: Batch delete schedule tasks
      tags:
      - auth required
  /batch-delete/ddns:
    post:
      consumes:
      - application/json
      description: Batch delete DDNS configurations
      parameters:
      - description: id list
        in: body
        name: request
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: Batch delete DDNS configurations
      tags:
      - auth required
  /batch-delete/nat:
    post:
      consumes:
      - application/json
      description: Batch delete NAT configurations
      parameters:
      - description: id list
        in: body
        name: request
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: Batch delete NAT configurations
      tags:
      - auth required
  /batch-delete/notification:
    post:
      consumes:
      - application/json
      description: Batch delete notifications
      parameters:
      - description: id list
        in: body
        name: request
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: Batch delete notifications
      tags:
      - auth required
  /batch-delete/notification-group:
    post:
      consumes:
      - application/json
      description: Batch delete notification group
      parameters:
      - description: id list
        in: body
        name: request
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: Batch delete notification group
      tags:
      - auth required
  /batch-delete/server:
    post:
      consumes:
      - application/json
      description: Batch delete server
      parameters:
      - description: id list
        in: body
        name: request
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: Batch delete server
      tags:
      - auth required
  /batch-delete/server-group:
    post:
      consumes:
      - application/json
      description: Batch delete server group
      parameters:
      - description: id list
        in: body
        name: request
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: Batch delete server group
      tags:
      - auth required
  /batch-delete/service:
    post:
      consumes:
      - application/json
      description: Batch delete service
      parameters:
      - description: id list
        in: body
        name: request
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: Batch delete service
      tags:
      - auth required
  /batch-delete/user:
    post:
      consumes:
      - application/json
      description: Batch delete users
      parameters:
      - description: id list
        in: body
        name: request
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: Batch delete users
      tags:
      - admin required
  /batch-delete/waf:
    patch:
      consumes:
      - application/json
      description: Edit server
      parameters:
      - description: block list
        in: body
        name: request
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: Edit server
      tags:
      - admin required
  /batch-move/server:
    post:
      consumes:
      - application/json
      description: Batch move servers to other user
      parameters:
      - description: BatchMoveServerForm
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.BatchMoveServerForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: Batch move servers to other user
      tags:
      - auth required
  /cron:
    get:
      description: List schedule tasks
      parameters:
      - description: Resource ID
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-array_model_Cron'
      security:
      - BearerAuth: []
      summary: List schedule tasks
      tags:
      - auth required
    post:
      consumes:
      - application/json
      description: Create new schedule task
      parameters:
      - description: CronForm
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CronForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-uint64'
      security:
      - BearerAuth: []
      summary: Create new schedule task
      tags:
      - auth required
  /cron/{id}:
    patch:
      consumes:
      - application/json
      description: Update schedule task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: CronForm
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CronForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: Update schedule task
      tags:
      - auth required
  /cron/{id}/manual:
    get:
      consumes:
      - application/json
      description: Trigger schedule task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: Trigger schedule task
      tags:
      - auth required
  /ddns:
    get:
      description: List DDNS profiles
      parameters:
      - description: Resource ID
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-array_model_DDNSProfile'
      security:
      - BearerAuth: []
      summary: List DDNS profiles
      tags:
      - auth required
    post:
      consumes:
      - application/json
      description: Add DDNS profile
      parameters:
      - description: DDNS Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.DDNSForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-uint64'
      security:
      - BearerAuth: []
      summary: Add DDNS profile
      tags:
      - auth required
  /ddns/{id}:
    patch:
      consumes:
      - application/json
      description: Edit DDNS profile
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: integer
      - description: DDNS Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.DDNSForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: Edit DDNS profile
      tags:
      - auth required
  /ddns/providers:
    get:
      description: List DDNS providers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-array_string'
      security:
      - BearerAuth: []
      summary: List DDNS providers
      tags:
      - auth required
  /file:
    get:
      consumes:
      - application/json
      description: Create an "attached" FM. It is advised to only call this within
        a terminal session.
      parameters:
      - description: Server ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CreateFMResponse'
      summary: Create FM session
      tags:
      - auth required
  /force-update/server:
    post:
      consumes:
      - application/json
      description: Force update Agent
      parameters:
      - description: id list
        in: body
        name: request
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-model_ServerTaskResponse'
      security:
      - BearerAuth: []
      summary: Force update Agent
      tags:
      - auth required
  /login:
    post:
      consumes:
      - application/json
      description: user login
      parameters:
      - description: Login Request
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/model.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-model_LoginResponse'
      summary: user login
  /nat:
    get:
      description: List NAT profiles
      parameters:
      - description: Resource ID
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-array_model_NAT'
      security:
      - BearerAuth: []
      summary: List NAT profiles
      tags:
      - auth required
    post:
      consumes:
      - application/json
      description: Add NAT profile
      parameters:
      - description: NAT Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.NATForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-uint64'
      security:
      - BearerAuth: []
      summary: Add NAT profile
      tags:
      - auth required
  /nat/{id}:
    patch:
      consumes:
      - application/json
      description: Edit NAT profile
      parameters:
      - description: Profile ID
        in: path
        name: id
        required: true
        type: integer
      - description: NAT Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.NATForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: Edit NAT profile
      tags:
      - auth required
  /notification:
    get:
      description: List notification
      parameters:
      - description: Resource ID
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-array_model_Notification'
      security:
      - BearerAuth: []
      summary: List notification
      tags:
      - auth required
    post:
      consumes:
      - application/json
      description: Add notification
      parameters:
      - description: NotificationForm
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.NotificationForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: Add notification
      tags:
      - auth required
  /notification-group:
    get:
      description: List notification group
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-array_model_NotificationGroupResponseItem'
      security:
      - BearerAuth: []
      summary: List notification group
      tags:
      - auth required
    post:
      consumes:
      - application/json
      description: New notification group
      parameters:
      - description: NotificationGroupForm
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.NotificationGroupForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: New notification group
      tags:
      - auth required
  /notification-group/{id}:
    patch:
      consumes:
      - application/json
      description: Edit notification group
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: NotificationGroupForm
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.NotificationGroupForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: Edit notification group
      tags:
      - auth required
  /notification/{id}:
    patch:
      consumes:
      - application/json
      description: Edit notification
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      - description: NotificationForm
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.NotificationForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.CommonResponse-any'
      security:
      - BearerAuth: []
      summary: Edit notification
      tags:
      - auth required
  /online-user:
    get:
      description: List online users
      parameters:
      - description: Page limit
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PaginatedResponse-array_model_OnlineUser-model_OnlineUser'
      security:
      - BearerAuth: []
      summary: List online users
      tags:
      - auth required
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"